web/sentence-api/requirements.txt
Flask
flask-restful
pymongo
flask-marshmallow
marshmallow 
bcrypt
Flask-Mail
pyjwt
python-dotenv